Week 8/Day 1/Application Start/requirements.txt
chainlit==0.7.700
httpx==0.24.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-huggingface==0.1.0
langchain-openai==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langsmith==0.1.121
numpy==1.26.4
protobuf==4.25.5
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.1.1
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
unstructured==0.15.7